vm: move custom opcode tests, add deep copy test
